requirements.txt
httpx
Pillow
PyMuPDF
moviepy
wheel
python-dotenv
torch
opencv-python
scikit-image
requests
google-generativeai

